+++ app/provincia.py
@@ -1,5 +1,5 @@
 #se importa flask
-from flask import Blueprint, render_template, request, url_for, redirect,flash
+from flask import Blueprint, render_template, request, url_for, redirect,flash, session
 #se importa db.py para utilizar la conexion a mysql
 from db import mysql
 #importamos el modulo que creamos
@@ -10,6 +10,9 @@ provincias = Blueprint('provincias', __name__, template_folder='app/templates')
 #envia los datos a la vista principalde provincia
 @provincias.route('/provincia')
 def Provincia():
+    if "user" not in session:
+        flash("you are NOT authorized")
+        return redirect("/ingresar")
     cur = mysql.connection.cursor()
     cur.execute('SELECT * from provincia')
     data = cur.fetchall()
@@ -18,6 +21,9 @@ def Provincia():
 #agrega un registro para provincia
 @provincias.route('/add_provincia', methods = ['POST'])
 def add_provincia():
+    if "user" not in session:
+        flash("you are NOT authorized")
+        return redirect("/ingresar")
     if request.method == 'POST':
         nombre_provincia = request.form['nombre_provincia']
         try:
@@ -33,6 +39,9 @@ def add_provincia():
 #enviar datos a vista editar segun el id
 @provincias.route('/edit_provincia/<id>', methods = ['POST', 'GET'])
 def edit_provincia(id):
+    if "user" not in session:
+        flash("you are NOT authorized")
+        return redirect("/ingresar")
     try:
         cur = mysql.connection.cursor()
         cur.execute('SELECT * FROM provincia WHERE idProvincia = %s', (id,))
@@ -45,6 +54,9 @@ def edit_provincia(id):
 #actualiza el registro segun id 
 @provincias.route('/update_provincia/<id>', methods = ['POST'])
 def update_provincia(id):
+    if "user" not in session:
+        flash("you are NOT authorized")
+        return redirect("/ingresar")
     if request.method == 'POST':
         nombre_provincia = request.form['nombre_provincia']
         try:
@@ -64,6 +76,9 @@ def update_provincia(id):
 #elimina un elemento segun el id correspondiente
 @provincias.route('/delete_provincia/<id>', methods = ['POST', 'GET'])
 def delete_provincia(id):
+    if "user" not in session:
+        flash("you are NOT authorized")
+        return redirect("/ingresar")
     try:
         cur = mysql.connection.cursor()
         cur.execute('DELETE FROM provincia WHERE idProvincia = %s', (id,))
